requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.0
schedule>=1.2.1
jinja2>=3.1.4
pyyaml>=6.0.1
python-dotenv>=1.0.1
rich>=13.7.0
tenacity>=8.2.3
fake-useragent>=1.5.1
tabulate>=0.9.0
